+++ app/pkg/const.go
@@ -1,0 +1,6 @@
+package pkg
+
+const (
+	ResponseOK = "OK"
+	ResponseNG = "NG"
+)
